+++ CHANGELOG.md
@@ -6,10 +6,15 @@ The following is a curated list of changes in the Enact agate module, newest cha
 
 ### Added
 
+- `agate/ContextualPopup` prop `css` to support customizations
+- `agate/ContextualPopupDecorator` prop `css` to support customizations
+- `agate/Dropdown` prop `width` to support multiple widths
+- `agate/Item` to have a `selected` publicClassname
 - `agate/Spinner` prop `children` to support the display of text below the icon
 
 ### Changed
 
+- `agate/Dropdown` to use `agate/ContextualPopup`
 - `agate/ThemeDecorator` to exclude `enact-fit` className when `disableFullscreen` is true
 
 ### Fixed